Initial working version
This is the initial version of a class with automatic data augmentation. It performs rotations, translations, and shearing to each image as it is generated for training. Unfortunately, it currently does this for validation images as well (this will need to be changed). There areno annotations to the parameters yet, but the code has docstrings. There is probably a more efficient way to do the looping (so its not done twice).